+++ demo/components/Validation/Form.tsx
@@ -3,7 +3,10 @@ import * as React from 'react';
 import * as L from '../../../leda';
 import { StoryProps } from '../../types';
 
-export const Form = (props: StoryProps) => (
+export const Form = (props: StoryProps) => {
+  const [value, setValue] = React.useState<string | number>(0);
+
+  return (
   <L.Div _box _inner>
     <L.Div>
       <L.Div _inner>
@@ -32,6 +35,29 @@ export const Form = (props: StoryProps) => (
         />
       </L.Div>
       <L.Div _inner>
+      <L.RadioGroup
+          form="form2"
+          name="radio"
+          value={value}
+          wrapperRender={({ elementProps }) => <L.Div {...elementProps} />}
+          onChange={ev => {
+            console.log('ev.component.value', ev.component.value);
+            console.log('ev.component.name', ev.component.name);
+            setValue(ev.component.value);
+          }}
+        >
+          <L.RadioButton 
+            value={0} 
+            wrapperRender={({ elementProps }) => 
+              <L.Span {...elementProps} style={{ color: 'green' }} 
+          />}>
+            form2 One
+            </L.RadioButton>
+          <L.RadioButton value={1}>form2 Two</L.RadioButton>
+          <L.RadioButton value={2}>form2 Three</L.RadioButton>
+        </L.RadioGroup>
+      </L.Div>
+      <L.Div _inner>
         <L.Input
           isRequired
           form="form2"
@@ -48,6 +74,7 @@ export const Form = (props: StoryProps) => (
         </L.Button>
         {' '}
         <L.Button
+          onClick={(ev) => console.log(ev.forms)}
           form="form2"
           _warning
         >
@@ -64,3 +91,4 @@ export const Form = (props: StoryProps) => (
     </L.Div>
   </L.Div>
 );
+        }