+++ leda/components/DatePicker/DatePicker.test.tsx
@@ -6,16 +6,20 @@ import {
 
 import userEvent from '@testing-library/user-event';
 import { DatePicker } from './index';
+import { MonthsNames, WeekDayNames } from '../../src/Calendar/types';
 
 const validName = 'test';
 const validFormat = 'dd.MM.yyyy';
 const invalidFormat = 'yyyy-MM-dd';
 const validValue = '10.10.2010';
 const invalidValue = '2010-10-10';
-
+const customMonthNames: MonthsNames = ['January', 'February', 'March', 'April', 'May', 'June', 'July', 'August', 'September', 'October', 'November', 'December'];
+const customShortMonthNames: MonthsNames = ['Jan', 'Feb', 'Mar', 'Apr', 'May', 'Jun', 'Jul', 'Aug', 'Sep', 'Oct', 'Nov', 'Dec'];
+const customWeekDayNames: WeekDayNames = ['Monday', 'Tuesday', 'Wednesday', 'Thursday', 'Friday', 'Saturday', 'Sunday'];
+const customShortWeekDayNames: WeekDayNames = ['Mon', 'Tue', 'Wed', 'Thu', 'Fri', 'Sat', 'Sun'];
 
 describe('DatePicker snapshots collection', () => {
-  test('is DatePicker render right?', () => {
+  test('is DatePicker rendering right?', () => {
     const { container } = render(<DatePicker name={validName} />);
 
     expect(container.firstChild)
@@ -33,7 +37,7 @@ describe('DatePicker attributes test collection', () => {
     expect(getByRole('textbox'))
       .toHaveProperty('placeholder', validPlaceholder);
   });
-  test('is Datepicker work right witn minMax set attributes?', () => {
+  test('is Datepicker working right with minMax set attributes?', () => {
     const min = new Date('01.02.2018');
     const max = new Date('05.25.2020');// mm-dd-YYYY
     const validDate = '25.05.2020';
@@ -43,7 +47,59 @@ describe('DatePicker attributes test collection', () => {
     expect(getByText(valueForCheck))
       .toHaveClass('calendar-date-cell disabled-date');
   });
-  test('is Datepicker work right with isDisabled attributes?', () => {
+  test('is Datepicker working right with max set attribute and onPressEnter event listener?', () => {
+    const max = new Date('05.25.2020');// mm-dd-YYYY
+    const invalidDate = '26.05.2020';
+    const validDate = '25.05.2020';
+    const onChange = jest.fn();
+    const { container } = render(<DatePicker name={validName} value={invalidDate} max={max} onChange={onChange} />);
+    const input = container.querySelectorAll('input.datepicker-input')[0];
+
+    fireEvent.keyDown(input, {
+      charCode: 13,
+      code: 13,
+      key: 'Enter',
+      keyCode: 13,
+    });
+
+    expect(onChange)
+      .toHaveBeenCalledTimes(1);
+
+    expect(onChange)
+      .lastCalledWith(expect.objectContaining({
+        component: expect.objectContaining({
+          name: validName,
+          value: validDate,
+        }),
+      }));
+  });
+  test('is Datepicker working right with min set attribute and onPressEnter event listener?', () => {
+    const min = new Date('01.02.2018');// mm-dd-YYYY
+    const invalidDate = '01.01.2018';
+    const validDate = '02.01.2018';
+    const onChange = jest.fn();
+    const { container } = render(<DatePicker name={validName} value={invalidDate} min={min} onChange={onChange} />);
+    const input = container.querySelectorAll('input.datepicker-input')[0];
+
+    fireEvent.keyDown(input, {
+      charCode: 13,
+      code: 13,
+      key: 'Enter',
+      keyCode: 13,
+    });
+
+    expect(onChange)
+      .toHaveBeenCalledTimes(1);
+
+    expect(onChange)
+      .lastCalledWith(expect.objectContaining({
+        component: expect.objectContaining({
+          name: validName,
+          value: validDate,
+        }),
+      }));
+  });
+  test('is Datepicker working right with isDisabled attributes?', () => {
     const onChange = jest.fn();
     const { container } = render(<DatePicker isDisabled value={validValue} onChange={onChange} />);
     const icon = container.querySelectorAll('.datepicker-icons-wrapper')[0];
@@ -60,7 +116,7 @@ describe('DatePicker attributes test collection', () => {
     expect(onChange)
       .toHaveBeenCalledTimes(0);
   });
-  test('is Datepicker work right with isOpen attribute?', () => {
+  test('is Datepicker working right with isOpen attribute?', () => {
     const onChange = jest.fn();
     const { container, rerender } = render(<DatePicker isOpen value={validValue} onChange={onChange} />);
     const icon = container.querySelectorAll('.datepicker-icons-wrapper')[0];
@@ -82,7 +138,7 @@ describe('DatePicker attributes test collection', () => {
       .not
       .toBeDefined();
   });
-  test('is Datepicker work right with date format input?', () => {
+  test('is Datepicker working right with date format input?', () => {
     const onChange = jest.fn();
     const { container, rerender } = render(<DatePicker format={validFormat} value={validValue} onChange={onChange} />);
     const icon = container.querySelectorAll('.datepicker-icons-wrapper')[0];
@@ -100,9 +156,72 @@ describe('DatePicker attributes test collection', () => {
     expect(input)
       .toHaveValue(invalidValue);
   });
+  test('is Datepicker working right with monthNames attribute?', () => {
+    const date = new Date();
+    const currentMonth = date.getMonth();
+    const { container } = render(<DatePicker monthNames={customMonthNames} isOpen />);
+    const calendar = container.querySelectorAll('.calendar-wrapper.visible')[0];
+    const title = container.querySelector('.calendar-title');
+    const titleMonth = title?.textContent?.split(' ')[0];
+
+    expect(calendar)
+      .toBeInTheDocument();
+
+    expect(title)
+      .toBeInTheDocument();
+
+    expect(customMonthNames).toContain(titleMonth);
+
+    expect(titleMonth).toBe(customMonthNames[currentMonth]);
+  });
+  test('is Datepicker working right with shortMonthNames attribute?', () => {
+    const { container } = render(<DatePicker shortMonthNames={customShortMonthNames} isOpen />);
+    const calendar = container.querySelectorAll('.calendar-wrapper.visible')[0];
+    const title = calendar.querySelector('.calendar-title');
+
+    fireEvent.click(title as Element);
+
+    const monthCells = container.querySelectorAll('.calendar-month-year-cell');
+
+    customShortMonthNames.forEach((monthName, index) => {
+      expect(monthCells[index])
+        .toBeInTheDocument();
+
+      expect(monthCells[index].textContent)
+        .toBe(monthName);
+    });
+  });
+  test('is Datepicker working right with weekDayNames attribute?', () => {
+    const { container } = render(<DatePicker weekDayNames={customWeekDayNames} isOpen />);
+    const calendar = container.querySelectorAll('.calendar-wrapper.visible')[0];
+    const weekDaysRow = calendar.querySelector('.calendar-week-days');
+    const weekDaysCells = weekDaysRow?.querySelectorAll('.calendar-date-cell') as NodeListOf<Element>;
+
+    customWeekDayNames.forEach((weekDayName, index) => {
+      expect(weekDaysCells[index])
+        .toBeInTheDocument();
+
+      expect(weekDaysCells[index].getAttribute('title'))
+        .toBe(weekDayName);
+    });
+  });
+  test('is Datepicker working right with shortWeekDayNames attribute?', () => {
+    const { container } = render(<DatePicker shortWeekDayNames={customShortWeekDayNames} isOpen />);
+    const calendar = container.querySelectorAll('.calendar-wrapper.visible')[0];
+    const weekDaysRow = calendar.querySelector('.calendar-week-days');
+    const weekDaysCells = weekDaysRow?.querySelectorAll('.calendar-date-cell') as NodeListOf<Element>;
+
+    customShortWeekDayNames.forEach((weekDayName, index) => {
+      expect(weekDaysCells[index])
+        .toBeInTheDocument();
+
+      expect(weekDaysCells[index].textContent)
+        .toBe(weekDayName);
+    });
+  });
 });
 describe('Datepicker event listeners test collection', () => {
-  test('is Datepicker work right with onBlur event listener?', () => {
+  test('is Datepicker working right with onBlur event listener?', () => {
     const onBlur = jest.fn();
     const { container } = render(<DatePicker value={validValue} name={validName} onBlur={onBlur} />);
     const input = container.querySelectorAll('input.datepicker-input')[0];
@@ -124,7 +243,7 @@ describe('Datepicker event listeners test collection', () => {
         }),
       }));
   });
-  test('is Datepicker work right with onChange event listener?', () => {
+  test('is Datepicker working right with onChange event listener?', () => {
     const validValueWithoutComma = '10101010';
     const onChange = jest.fn();
     const { container } = render(<DatePicker value={invalidValue} format={validFormat} name={validName} onChange={onChange} />);
@@ -143,7 +262,7 @@ describe('Datepicker event listeners test collection', () => {
         }),
       }));
   });
-  test('is Datepicker work right with onPressEnter event listener?', () => {
+  test('is Datepicker working right with onPressEnter event listener?', () => {
     const onEnterPress = jest.fn();
     const { container } = render(<DatePicker name={validName} value={validValue} onEnterPress={onEnterPress} />);
     const input = container.querySelectorAll('input.datepicker-input')[0];
@@ -166,7 +285,7 @@ describe('Datepicker event listeners test collection', () => {
         }),
       }));
   });
-  test('is Datepicker work right with onFocus event listener?', () => {
+  test('is Datepicker working right with onFocus event listener?', () => {
     const onFocus = jest.fn();
     const { container } = render(<DatePicker name={validName} value={validValue} onFocus={onFocus} />);
     const input = container.querySelectorAll('input.datepicker-input')[0];
@@ -186,7 +305,7 @@ describe('Datepicker event listeners test collection', () => {
   });
 });
 describe('DatePicker quality set test collection', () => {
-  test('is DatePicker render right if value set as String', () => {
+  test('is DatePicker rendering right if value set as String', () => {
     const dateValue = '10.10.2010';
     const {
       getAllByRole,
@@ -202,7 +321,7 @@ describe('DatePicker quality set test collection', () => {
     expect(getByRole('textbox'))
       .toHaveValue(dateValue);
   });
-  test('is DatePicker render right if value set as Date', () => {
+  test('is DatePicker rendering right if value set as Date', () => {
     const dateValue = new Date('10.10.2010');
     const {
       getAllByRole,
@@ -218,7 +337,7 @@ describe('DatePicker quality set test collection', () => {
     expect(getByRole('textbox'))
       .toHaveValue('10.10.2010');
   });
-  test('is DatePicker render right if value set as Null', () => {
+  test('is DatePicker rendering right if value set as Null', () => {
     const dateValue = null;
     const {
       getAllByRole,